Cambios a nivel domain: EntregaDelExamen contiene el id de inscripcion del estudiante, no el id DEL estudiante.Ademas, creamos el repositorio de las inscripciones y por otra parte la entidad curso carga examenes recibiendo la entidad examen .  A nivel application: Desarrollamos handlers de poner nota en la entrega del examen y subir el modelo del examen, faltaria subir la entrega del examen

+++ Application/Commands/Examenes/PonerNotaEntregaExamenCommand.cs
@@ -3,4 +3,4 @@ using PlataformaDeGestionDeCursosOnline.Domain.Abstractions;
 
 namespace PlataformaDeGestionDeCursosOnline.Application.Commands.Examenes;
 
-public record PonerNotaEntregaExamenCommand(Guid IdEntregaExamen, double NuevaNota) : ICommand<Result>;
+public record PonerNotaEntregaExamenCommand(Guid IdEntregaExamen, decimal NuevaNota) : ICommand<Result>;
